chore(pages): move from app dir to pages
