+++ next.config.mjs
@@ -17,9 +17,6 @@ const config = {
   },
   reactStrictMode: true,
   swcMinify: true,
-  experimental: {
-    appDir: true,
-  },
   i18n: {
     locales: ["en"],
     defaultLocale: "en",
