--- src/app/(landing)/head.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const LandingHead = () => {
-  return (
-    <>
-      <title>Broaden Agency</title>
-      <meta name="description" content="Broaden Agency" />
-      <link rel="icon" href="/favicon.ico" />
-    </>
-  );
-};
-
-export default LandingHead;
