--- src/app/(landing)/page.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const LandingPage = () => {
-  return <div></div>;
-};
-
-export default LandingPage;
