+++ src/pages/api/auth/[...nextauth].ts
@@ -1,3 +1,4 @@
+// Prisma adapter for NextAuth, optional and can be removed
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import NextAuth, { type NextAuthOptions } from "next-auth";
