+++ src/pages/index.tsx
@@ -1,0 +1,20 @@
+import { type NextPage } from "next";
+import Head from "next/head";
+import MainContainer from "src/components/container";
+
+const Home: NextPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Broaden Agency</title>
+        <meta name="description" content="Generated by create-t3-app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <MainContainer>
+        <main></main>
+      </MainContainer>
+    </>
+  );
+};
+
+export default Home;
